Fix spring boot starter failing without logback
